+++ src/pages/sections/about-us/FeatureSection.tsx
@@ -1,66 +1,7 @@
 import { forwardRef, useEffect, useRef, useState } from 'react';
+import { featureData } from '@/lib/whoweareData';
 
-import s1Img from '@/asset/imgs/about-us-Img/s1Img.jpg';
-import s2Img from '@/asset/imgs/about-us-Img/s2Img.jpg';
-import whoweare_b from '@/asset/imgs/about-us-Img/whoweare_b.jpg';
-import whoweare_miro from '@/asset/imgs/about-us-Img/whoweare_miro.jpg';
-
-interface Feature {
-    title: string;
-    heading: string;
-    description: string;
-    listItems: string[];
-    imageSrc: string;
-    bgColor: string;
-    primaryTextColor: string;
-    secondaryTextColor: string;
-}
-
-const featureData: Feature[] = [
-    {
-        title: "Our Mission",
-        heading: "iGaming을 통해\n세상을 연결합니다",
-        description: "NSUS Group은 현재 북미, 유럽, 아시아 등 지역에 700명 이상의 임직원들이 일하고 있는 다국적 기업으로 iGaming 업계의 유니콘이며, NSUSLAB은 NSUS Group의 개발 스튜디오로서 그룹의 핵심적인 역할을 담당하고 있습니다.\n\n우리는 iGaming을 통해 세상을 연결하고자 합니다. 이를 통해 유저의 즐거움을 극대화 하는 것이 NSUSLAB이 지향하는 궁극적인 목표이자 존재의 이유입니다.",
-        listItems: ["NSUS Group - Forefront of marketing and technology", "NSUSLAB - Development Studio"],
-        imageSrc: s1Img,
-        bgColor: '#111111',
-        primaryTextColor: '#ffffff',
-        secondaryTextColor: '#ffffff',
-    },
-    {
-        title: "Our Vision",
-        heading: "Global No.1\niGaming Company",
-        description: "NSUSLAB은 그 누구도 감히 상상하지 못했던 Global No.1 Poker Game 이라는 성공 신화를 만든 노하우를 바탕으로 Casino 시장에서도 또 한 번 혁신적인 제품을 가지고 세계 최고가 되고자 합니다. \n\n이를 통해 iGaming Industry 전반에서 독보적인 리더로 자리매김할 것입니다.",
-        listItems: ["GGPoker - Classic Games", "GGPoker - Tournament", "GGVegas - Table Games", "GGVegas - Slots","ClubGG","WSOP+" ],
-        imageSrc: s2Img,
-        bgColor: '#841e09',
-        primaryTextColor: '#ffffff',
-        secondaryTextColor: '#ffffff',
-    },
-    {
-        title: "Our Value",
-        heading: "새로운 차원의\n엔터테인먼트를 제공합니다",
-        description: "NSUSLAB은 새로운 차원의 엔터테인먼트를 제공하여 \nGlobal iGaming Community의 리더가 되고자 합니다. \n\n명품을 만들기 위해서는 끊임없는 노력과 인내가 필요합니다. 이 도전이 쉽지 않은 것을 알기에 확신이 없는 프로젝트는 시작하지 않을 것이며, 한번 시작한 프로젝트는 명품이 될 때까지 지속적인 지원이 이루어 지도록 할 것입니다.",
-        listItems: [],
-        imageSrc: whoweare_b,
-        bgColor: '#4f46e5',
-        primaryTextColor: '#ffffff',
-        secondaryTextColor: '#ffedd5',
-    },
-    {
-        title: "We're Hiring",
-        heading: "NSUSLAB의 위대한 도전을\n함께하실 분을 찾습니다.",
-        description: "외부 투자에 의지하고 성공에 대한 막연한 희망으로 프로젝트를 진행하는 대부분의 게임 개발사와는 달리, NSUSLAB은 성공한 제품과 사업에 의한 안정적 매출을 기반으로 보다 큰 성공을 이루기 위해 달리고 있습니다. \n\n 우리는 단지 꿈만 꾸는 것이 아닌, 좋은 현실을 더 좋게 만들고자 합니다. 안정적인 개발 환경에서 세계 시장을 제패하기 위해 도전적 성장과 경험을 함께 할 좋은 인재들을 적극 영입하고 있습니다.",
-        listItems: ["도전과 성장의 기회", "자유롭고 수평적인 문화", "최고의 보상과 근무 환경"],
-        imageSrc: whoweare_miro,
-        bgColor: '#111111',
-        primaryTextColor: '#ffffff',
-        secondaryTextColor: '#ffffff',
-    }
-];
-
-
-const FeatureSection = forwardRef<HTMLElement, {}>((props, ref) => {
+const FeatureSection = forwardRef<HTMLElement, {}>((_props, ref) => {
     const [activeIndex, setActiveIndex] = useState(0);
     const textSectionRefs = useRef<(HTMLDivElement | null)[]>([]);
     const imageRefs = useRef<(HTMLImageElement | null)[]>([]);
@@ -85,7 +26,6 @@ const FeatureSection = forwardRef<HTMLElement, {}>((props, ref) => {
 
         const update = () => {
             const vh = window.innerHeight;
-            // 텍스트 사라지는 시점 : fadeOutLine이 낮을수록 늦게 낮아짐
             const fadeOutLine = vh * 0.30; 
             const centerLine = vh * 0.50;
             const preInLine = vh * 3.0;
